Addressing review feedback.
